+++ spec/rails_ai_build/tasks/queue_spec.rb
@@ -11,11 +11,13 @@ RSpec.describe RailsAiBuild::Tasks::Queue do
   end
 
   before do
-    described_class.reset!
+    RailsAiBuild::Tasks::Queue.reset!
     RailsAiBuild.configure do |c|
       c.sync_tasks = true
       c.multitask_enabled = true
       c.api_keys[:openai] = 'sk-test'
+      c.branch_per_task = false
+      c.auto_pr_on_complete = false
     end
     allow_any_instance_of(RailsAiBuild::Tasks::Runtime).to receive(:run!).and_return(success_result)
   end